+++ pkg/codegen/schema.go
@@ -140,7 +140,8 @@ func AddType(schema map[string]any, name string, types map[string]pschema.Comple
 	propertySpecs := map[string]pschema.PropertySpec{}
 	for propertyName := range properties {
 		// Ignore unnamed properties like "-".
-		if strcase.ToCamel(propertyName) == "" {
+		camelCase := strcase.ToCamel(propertyName)
+		if camelCase == "" {
 			continue
 		}
 		propertySchema, _, _ := unstructured.NestedMap(properties, propertyName)